pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Chamoda Pandithage <chamoda@xaventra.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = {extras = ["email"], version = "^2.10.3"}
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
alembic = "^1.13.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
typer = "^0.12.1"
fastapi-pagination = "^0.12.22"
asyncpg = "^0.29.0"
pydantic-extra-types = "^2.7.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pyinfra = "^2.9.2"

[tool.pyright]
include = ["app"]
ignore = ["migrations/*"]
venvPath = "."
venv = "venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
